requirements.txt
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi
unicorn
groq